refactor: move findPatientCommand from parser to commands package
