+++ src/test/java/seedu/address/logic/commands/patient/FindPatientCommandTest.java
@@ -17,7 +17,7 @@ import java.util.Collections;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.parser.patient.FindPatientCommand;
+import seedu.address.logic.commands.patient.FindPatientCommand;
 import seedu.address.model.AddressBook;
 import seedu.address.model.AppointmentSchedule;
 import seedu.address.model.Model;
